requirements.txt
arrow==1.3.0
certifi==2025.4.26
cffi==1.17.1
circuitbreaker==2.1.3
click==8.0.4
cryptography==44.0.3
decorator==5.2.1
greenlet==3.2.2
iniconfig==2.1.0
jmespath==0.10.0
markdown-it-py==3.0.0
mdurl==0.1.2
oci==2.153.0
oci-cli==3.57.0
packaging==25.0
pluggy==1.6.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
pyOpenSSL==24.3.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.1
rich==14.0.0
self==2020.12.3
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.41
terminaltables==3.1.10
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
wcwidth==0.2.13
